Reorganize HTTP endpoints
