+++ src/servers/http_server.h
@@ -25,39 +25,284 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
+#include <re2/re2.h>
+#include <list>
 #include <map>
 #include <memory>
 #include <string>
+#include <thread>
+#include <unordered_map>
+#include "src/core/constants.h"
+#include "src/core/logging.h"
+#include "src/servers/common.h"
+#include "src/servers/data_compressor.h"
 #include "src/servers/shared_memory_manager.h"
 #include "src/servers/tracer.h"
 #include "triton/core/tritonserver.h"
 
+#include <evhtp/evhtp.h>
+
 namespace nvidia { namespace inferenceserver {
 
+// Generic HTTP server using evhtp
 class HTTPServer {
  public:
-  static TRITONSERVER_Error* CreateAPIServer(
-      const std::shared_ptr<TRITONSERVER_Server>& server,
-      nvidia::inferenceserver::TraceManager* trace_manager,
-      const std::shared_ptr<SharedMemoryManager>& smb_manager,
-      const int32_t port, const int thread_cnt,
-      std::unique_ptr<HTTPServer>* http_server);
+  virtual ~HTTPServer() { IGNORE_ERR(Stop()); }
 
-  static TRITONSERVER_Error* CreateMetricsServer(
+  TRITONSERVER_Error* Start();
+  TRITONSERVER_Error* Stop();
+
+ protected:
+  explicit HTTPServer(const int32_t port, const int thread_cnt)
+      : port_(port), thread_cnt_(thread_cnt)
+  {
+  }
+
+
+  static void Dispatch(evhtp_request_t* req, void* arg);
+
+ protected:
+  virtual void Handle(evhtp_request_t* req) = 0;
+
+  static void StopCallback(int sock, short events, void* arg);
+
+  int32_t port_;
+  int thread_cnt_;
+
+  evhtp_t* htp_;
+  struct event_base* evbase_;
+  std::thread worker_;
+  int fds_[2];
+  event* break_ev_;
+};
+
+#ifdef TRITON_ENABLE_METRICS
+// Handle HTTP requests to obtain prometheus metrics
+class HTTPMetricsServer : public HTTPServer {
+ public:
+  static TRITONSERVER_Error* Create(
       const std::shared_ptr<TRITONSERVER_Server>& server, int32_t port,
       int thread_cnt, std::unique_ptr<HTTPServer>* metrics_server);
 
-  static TRITONSERVER_Error* CreateSagemakerAPIServer(
+  ~HTTPMetricsServer() = default;
+
+ private:
+  explicit HTTPMetricsServer(
+      const std::shared_ptr<TRITONSERVER_Server>& server, const int32_t port,
+      const int thread_cnt)
+      : HTTPServer(port, thread_cnt), server_(server),
+        api_regex_(R"(/metrics/?)")
+  {
+  }
+  void Handle(evhtp_request_t* req) override;
+
+  std::shared_ptr<TRITONSERVER_Server> server_;
+  re2::RE2 api_regex_;
+};
+#endif  // TRITON_ENABLE_METRICS
+
+// HTTP API server that implements KFServing community standard inference
+// protocols and extensions used by Triton.
+class HTTPAPIServer : public HTTPServer {
+ public:
+  static TRITONSERVER_Error* Create(
       const std::shared_ptr<TRITONSERVER_Server>& server,
       nvidia::inferenceserver::TraceManager* trace_manager,
       const std::shared_ptr<SharedMemoryManager>& smb_manager,
       const int32_t port, const int thread_cnt,
       std::unique_ptr<HTTPServer>* http_server);
 
-  virtual ~HTTPServer() = default;
+  virtual ~HTTPAPIServer();
 
-  virtual TRITONSERVER_Error* Start() = 0;
-  virtual TRITONSERVER_Error* Stop() = 0;
+  //
+  // AllocPayload
+  //
+  // Simple structure that carries the userp payload needed for
+  // allocation.
+  struct AllocPayload {
+    struct OutputInfo {
+      enum Kind { JSON, BINARY, SHM };
+      Kind kind_;
+
+      ~OutputInfo()
+      {
+        if (evbuffer_ != nullptr) {
+          evbuffer_free(evbuffer_);
+        }
+      }
+
+      // For shared memory
+      OutputInfo(void* b, uint64_t s, TRITONSERVER_MemoryType m, int64_t i)
+          : kind_(SHM), base_(b), byte_size_(s), memory_type_(m), device_id_(i),
+            evbuffer_(nullptr)
+      {
+      }
+      void* base_;
+      uint64_t byte_size_;
+      TRITONSERVER_MemoryType memory_type_;
+      int64_t device_id_;
+
+      // For non-shared memory
+      OutputInfo(Kind k, uint32_t class_cnt)
+          : kind_(k), class_cnt_(class_cnt), evbuffer_(nullptr)
+      {
+      }
+      uint32_t class_cnt_;
+      evbuffer* evbuffer_;
+    };
+
+    ~AllocPayload()
+    {
+      for (auto it : output_map_) {
+        delete it.second;
+      }
+    }
+
+    AllocPayload() : default_output_kind_(OutputInfo::Kind::JSON){};
+    std::unordered_map<std::string, OutputInfo*> output_map_;
+    AllocPayload::OutputInfo::Kind default_output_kind_;
+  };
+
+  // Object associated with an inference request. This persists
+  // information needed for the request and records the evhtp thread
+  // that is bound to the request. This same thread must be used to
+  // send the response.
+  class InferRequestClass {
+   public:
+    explicit InferRequestClass(
+        TRITONSERVER_Server* server, evhtp_request_t* req,
+        DataCompressor::Type response_compression_type);
+    virtual ~InferRequestClass() = default;
+
+    evhtp_request_t* EvHtpRequest() const { return req_; }
+
+    static void InferRequestComplete(
+        TRITONSERVER_InferenceRequest* request, const uint32_t flags,
+        void* userp);
+    static void InferResponseComplete(
+        TRITONSERVER_InferenceResponse* response, const uint32_t flags,
+        void* userp);
+    TRITONSERVER_Error* FinalizeResponse(
+        TRITONSERVER_InferenceResponse* response);
+
+    // Helper function to set infer response header in the form specified by
+    // the endpoint protocol
+    virtual void SetResponseHeader(
+        const bool has_binary_data, const size_t header_length);
+
+    uint32_t IncrementResponseCount();
+
+#ifdef TRITON_ENABLE_TRACING
+    TraceManager* trace_manager_;
+    uint64_t trace_id_;
+#endif  // TRITON_ENABLE_TRACING
+
+    AllocPayload alloc_payload_;
+
+    // Data that cannot be used directly from the HTTP body is first
+    // serialized. Hold that data here so that its lifetime spans the
+    // lifetime of the request.
+    std::list<std::vector<char>> serialized_data_;
+
+   protected:
+    TRITONSERVER_Server* server_;
+    evhtp_request_t* req_;
+    evthr_t* thread_;
+
+    DataCompressor::Type response_compression_type_;
+
+    // Counter to keep track of number of responses generated.
+    std::atomic<uint32_t> response_count_;
+  };
+
+ protected:
+  explicit HTTPAPIServer(
+      const std::shared_ptr<TRITONSERVER_Server>& server,
+      nvidia::inferenceserver::TraceManager* trace_manager,
+      const std::shared_ptr<SharedMemoryManager>& shm_manager,
+      const int32_t port, const int thread_cnt);
+  virtual void Handle(evhtp_request_t* req) override;
+  virtual std::unique_ptr<InferRequestClass> CreateInferRequest(
+      evhtp_request_t* req)
+  {
+    return std::unique_ptr<InferRequestClass>(new InferRequestClass(
+        server_.get(), req, GetResponseCompressionType(req)));
+  }
+
+  // Helper function to retrieve infer request header in the form specified by
+  // the endpoint protocol
+  //
+  // Get the inference header length. Return 0 if the whole request body is
+  // the inference header.
+  virtual size_t GetInferenceHeaderLength(evhtp_request_t* req);
+  virtual DataCompressor::Type GetRequestCompressionType(evhtp_request_t* req);
+  virtual DataCompressor::Type GetResponseCompressionType(evhtp_request_t* req);
+
+  static TRITONSERVER_Error* InferResponseAlloc(
+      TRITONSERVER_ResponseAllocator* allocator, const char* tensor_name,
+      size_t byte_size, TRITONSERVER_MemoryType preferred_memory_type,
+      int64_t preferred_memory_type_id, void* userp, void** buffer,
+      void** buffer_userp, TRITONSERVER_MemoryType* actual_memory_type,
+      int64_t* actual_memory_type_id);
+  static TRITONSERVER_Error* InferResponseFree(
+      TRITONSERVER_ResponseAllocator* allocator, void* buffer,
+      void* buffer_userp, size_t byte_size, TRITONSERVER_MemoryType memory_type,
+      int64_t memory_type_id);
+  void HandleServerHealth(evhtp_request_t* req, const std::string& kind);
+  void HandleServerMetadata(evhtp_request_t* req);
+  void HandleModelReady(
+      evhtp_request_t* req, const std::string& model_name,
+      const std::string& model_version_str);
+  void HandleModelMetadata(
+      evhtp_request_t* req, const std::string& model_name,
+      const std::string& model_version_str);
+  void HandleModelConfig(
+      evhtp_request_t* req, const std::string& model_name,
+      const std::string& model_version_str);
+  void HandleInfer(
+      evhtp_request_t* req, const std::string& model_name,
+      const std::string& model_version_str);
+  void HandleModelStats(
+      evhtp_request_t* req, const std::string& model_name = "",
+      const std::string& model_version_str = "");
+  void HandleRepositoryIndex(
+      evhtp_request_t* req, const std::string& repository_name);
+  void HandleRepositoryControl(
+      evhtp_request_t* req, const std::string& repository_name,
+      const std::string& model_name, const std::string& action);
+  void HandleSystemSharedMemory(
+      evhtp_request_t* req, const std::string& region_name,
+      const std::string& action);
+  void HandleCudaSharedMemory(
+      evhtp_request_t* req, const std::string& region_name,
+      const std::string& action);
+
+  TRITONSERVER_Error* EVBufferToInput(
+      const std::string& model_name, TRITONSERVER_InferenceRequest* irequest,
+      evbuffer* input_buffer, InferRequestClass* infer_req,
+      size_t header_length);
+
+  static void OKReplyCallback(evthr_t* thr, void* arg, void* shared);
+  static void BADReplyCallback(evthr_t* thr, void* arg, void* shared);
+
+  std::shared_ptr<TRITONSERVER_Server> server_;
+
+  // Storing server metadata as it is consistent during server running
+  TRITONSERVER_Error* server_metadata_err_;
+  std::string server_metadata_;
+
+  TraceManager* trace_manager_;
+  std::shared_ptr<SharedMemoryManager> shm_manager_;
+
+  // The allocator that will be used to allocate buffers for the
+  // inference result tensors.
+  TRITONSERVER_ResponseAllocator* allocator_;
+
+  re2::RE2 server_regex_;
+  re2::RE2 model_regex_;
+  re2::RE2 modelcontrol_regex_;
+  re2::RE2 systemsharedmemory_regex_;
+  re2::RE2 cudasharedmemory_regex_;
 };
 
 }}  // namespace nvidia::inferenceserver
